Changed external ID to consider bigints

+++ db/migrate/20180309223954_create_users.rb
@@ -3,9 +3,9 @@ class CreateUsers < ActiveRecord::Migration[5.0]
     create_table :users do |t|
       t.string :first_name
       t.string :last_name
-      t.integer :external_id, null: false
-      t.integer :sponsor_external_id, null: false
-      t.integer :placement_external_id, null: false
+      t.bigint :external_id, null: false
+      t.bigint :sponsor_external_id, null: false
+      t.bigint :placement_external_id, null: false
       t.string :phone
       t.boolean :active, default: true
